requirements.txt
Django
djangorestframework-gis
djangorestframework
